Refactored to seperate git header code from raml-design fs code.
The original raml-designer-4-git.html file now only contains the header
related code that allows to navigate the different tags and branches
related to the requested git repo.
A filter mechanism has been added when loading the tags and branches
which allows to configure what options are being provided for
navigation. It uses RegExp defined in a config.json file under the same
directory with name repo.replace(/\//, '_')+".json";
The code with the raml-designer-git-fs.html is now under a separate file
and gets loaded via a iframe to have clean code separation.